semana02/dia5/requirements.txt
# pyrebase : para integrar python con firebase
pyrebase4
#flask : para implementar web con python  utilizamos el microframework flask
Flask
#flask-bootstrap : permite integrar flask con bootstrap
flask-bootstrap
#Flask-WTF : permite trabajar con formularios en flask
Flask-WTF
#python-decouple : sirve para poder utilizar variables de entorno
python-decouple
#python-dotenv : sirve para poder utilizar variables de entorno
python-dotenv
#firebase-admin : sirve para poder trabajar con firestore
firebase-admin

